apple count
