+++ main/2/.travis.yml
@@ -1,0 +1,4 @@
+language: java
+jdk: openjdk8
+after_success:
+  - mvn coveralls:report
