+++ main/2/src/main/java/3/ApplesCount.java
@@ -1,0 +1,10 @@
+//package 1;
+import java.util.Scanner;
+public class ApplesCount {
+    int applenum() {
+        Scanner scanner = new Scanner(System.in);
+        System.out.print("りんごいくつ買いますか？: ");
+        int num2 = scanner.nextInt();
+        return num2;
+    }
+}
