+++ main/2/src/main/java/3/CaluculateApple.java
@@ -1,0 +1,10 @@
+//package 3;
+
+public class CaluculateApple {
+    int applesumvalue(int sumapple){
+        int applevalue = 100;
+        int sumvalue = sumapple * applevalue;
+        System.out.println("りんご" + sumapple + "個の値段は" + sumvalue + "円です。");
+        return sumvalue;
+    }
+}
