+++ main/2/src/main/java/3/Main.java
@@ -1,0 +1,11 @@
+//import java.util.Scanner;
+//package 1;
+public class Main {
+    public static void main(String[] args) {
+        CaluculateApple calapp = new CaluculateApple();
+        ApplesCount appc = new ApplesCount();
+        //Recieved100 re100 = new Recieved100();
+        //re100.buy_apple();
+        calapp.applesumvalue(appc.applenum());
+    }
+}
