+++ main/2/src/test/java/3/ApplesTest.java
@@ -1,0 +1,12 @@
+package 1;
+import static org.junit.Assert.*;
+import org.junit.Test;
+
+public class ApplesTest {
+    @Test 
+    public void AppleCountTest() throws Exception {
+        ApplesCount AC = new ApplesCount();
+        int expected = 3;
+        assertThat(AC.applenum(),is(expected));
+    }
+}
